pyproject.toml
[tool.poetry]
name = "tickmate-connectordb"
version = "0.1.0"
description = "Upload your Tickmate data to ConnectorDB"
authors = ["Robbie Huffman <robbie.huffman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.5"
connectordb = "^0.3.5"
configloader = "^1.0"
pyyaml = "^3.13"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
